constants shared between components moved to the shared src/constants module
