+++ src/components/DecisioningEngine/consequenceAdapters/inAppMessageConsequenceAdapter.js
@@ -9,8 +9,8 @@ the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTA
 OF ANY KIND, either express or implied. See the License for the specific language
 governing permissions and limitations under the License.
 */
-import { MESSAGE_IN_APP } from "../../Personalization/constants/schema";
-import { TEXT_HTML } from "../../Personalization/constants/contentType";
+import { MESSAGE_IN_APP } from "../../../constants/schema";
+import { TEXT_HTML } from "../../../constants/contentType";
 
 export default (id, type, detail) => {
   const { html, mobileParameters } = detail;
