+++ src/components/Personalization/createPersonalizationDetails.js
@@ -22,7 +22,7 @@ import {
   REDIRECT_ITEM,
   RULESET_ITEM,
   MESSAGE_FEED_ITEM
-} from "./constants/schema";
+} from "../../constants/schema";
 
 const addPageWideScope = scopes => {
   if (!includes(scopes, PAGE_WIDE_SCOPE)) {
