+++ src/components/Personalization/createPreprocessors.js
@@ -9,7 +9,7 @@ the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTA
 OF ANY KIND, either express or implied. See the License for the specific language
 governing permissions and limitations under the License.
 */
-import { DOM_ACTION } from "./constants/schema";
+import { DOM_ACTION } from "../../constants/schema";
 import remapHeadOffers from "./dom-actions/remapHeadOffers";
 import remapCustomCodeOffers from "./dom-actions/remapCustomCodeOffers";
