+++ src/components/Personalization/createViewCacheManager.js
@@ -12,7 +12,7 @@ governing permissions and limitations under the License.
 
 import { assign, groupBy } from "../../utils";
 import defer from "../../utils/defer";
-import { DEFAULT_CONTENT_ITEM } from "./constants/schema";
+import { DEFAULT_CONTENT_ITEM } from "../../constants/schema";
 
 export default ({ createProposition }) => {
   const viewStorage = {};
