+++ src/components/Personalization/handlers/createProcessInAppMessage.js
@@ -9,7 +9,7 @@ the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTA
 OF ANY KIND, either express or implied. See the License for the specific language
 governing permissions and limitations under the License.
 */
-import { APPLICATION_JSON } from "../constants/contentType";
+import { APPLICATION_JSON } from "../../../constants/contentType";
 
 const DEFAULT_CONTENT = "defaultContent";
