+++ src/components/Personalization/index.js
@@ -32,7 +32,7 @@ import createPreprocess from "./dom-actions/createPreprocess";
 import injectCreateProposition from "./handlers/injectCreateProposition";
 import createAsyncArray from "./utils/createAsyncArray";
 import createPendingNotificationsHandler from "./createPendingNotificationsHandler";
-import * as schema from "./constants/schema";
+import * as schema from "../../constants/schema";
 import processDefaultContent from "./handlers/processDefaultContent";
 import { isPageWideSurface } from "./utils/surfaceUtils";
 import createProcessDomAction from "./handlers/createProcessDomAction";
