+++ test/unit/specs/components/Personalization/topLevel/buildAlloy.js
@@ -31,7 +31,7 @@ import injectCreateProposition from "../../../../../../src/components/Personaliz
 import createProcessPropositions from "../../../../../../src/components/Personalization/handlers/createProcessPropositions";
 import createAsyncArray from "../../../../../../src/components/Personalization/utils/createAsyncArray";
 import createPendingNotificationsHandler from "../../../../../../src/components/Personalization/createPendingNotificationsHandler";
-import * as schema from "../../../../../../src/components/Personalization/constants/schema";
+import * as schema from "../../../../../../src/constants/schema";
 import createProcessDomAction from "../../../../../../src/components/Personalization/handlers/createProcessDomAction";
 import createProcessHtmlContent from "../../../../../../src/components/Personalization/handlers/createProcessHtmlContent";
 import createProcessRedirect from "../../../../../../src/components/Personalization/handlers/createProcessRedirect";
